ama/requirements.txt
alembic==1.13.3
asgiref==3.8.1
asttokens==2.4.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
cryptography==43.0.3
debugpy==1.8.7
decorator==5.1.1
decouple==0.0.7
defusedxml==0.8.0rc2
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==65.1.0
django-cors-headers==4.5.0
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
executing==2.1.0
greenlet==3.1.1
idna==3.10
inflection==0.5.1
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
lightgbm==4.5.0
Mako==1.3.6
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.2
oauthlib==3.2.2
optuna==4.0.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
psutil==6.1.0
psycopg==3.2.3
psycopg2==2.9.10
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
scikit-learn==1.5.2
scipy==1.14.1
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
SQLAlchemy==2.0.36
sqlparse==0.5.1
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.1
tqdm==4.66.6
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
xgboost==2.1.2
